+++ widgets/views/toggle.php
@@ -21,12 +21,3 @@ use yii\helpers\Html;
         <span class="vote-count"><?= $count ?></span>
     </button>
 </div>
-
-<!--<div itemprop="aggregateRating" itemscope itemtype="http://schema.org/AggregateRating">-->
-<!--    <meta itemprop="interactionCount" content="Positive: --><?//= $positive?><!--"/>-->
-<!--    <meta itemprop="interactionCount" content="Negative: --><?//= $negative ?><!--"/>-->
-<!--    <meta itemprop="ratingValue" content="--><?//= $rating?><!--"/>-->
-<!--    <meta itemprop="ratingCount" content="--><?//= $positive + $negative?><!--"/>-->
-<!--    <meta itemprop="bestRating" content="10"/>-->
-<!--    <meta itemprop="worstRating" content="0"/>-->
-<!--</div>-->