+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [0.2.0-dev] - 2016-07-07
+### Added
+- Model and Query behaviors
+- Added `Favorite` widget
+
 ## [0.1.0] - 2016-07-02
 ### Added
 - Initial release